# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * E-Commerce API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface RegisterInput {
  name: string;
  email: string;
  password: string;
}

export interface LoginInput {
  email: string;
  password: string;
}

export type UserRole = (typeof UserRole)[keyof typeof UserRole];

export const UserRole = {
  customer: "customer",
  admin: "admin",
} as const;

export interface User {
  id: number;
  name: string;
  email: string;
  role: UserRole;
}

export interface AuthResponse {
  token: string;
  user: User;
}

export interface Product {
  id: number;
  name: string;
  description: string;
  price: number;
  image: string;
  category: string;
  countInStock: number;
  createdAt: string;
}

export interface ProductInput {
  name: string;
  description: string;
  price: number;
  image: string;
  category: string;
  countInStock: number;
}

export interface OrderItem {
  productId: number;
  name: string;
  price: number;
  quantity: number;
  image: string;
}

export type OrderStatus = (typeof OrderStatus)[keyof typeof OrderStatus];

export const OrderStatus = {
  pending: "pending",
  processing: "processing",
  shipped: "shipped",
  delivered: "delivered",
  cancelled: "cancelled",
} as const;

export interface Order {
  id: number;
  userId: number;
  /** @nullable */
  userName?: string | null;
  items: OrderItem[];
  totalPrice: number;
  status: OrderStatus;
  shippingAddress: string;
  createdAt: string;
}

export type OrderInputItemsItem = {
  productId: number;
  quantity: number;
};

export interface OrderInput {
  items: OrderInputItemsItem[];
  shippingAddress: string;
}

export type OrderStatusInputStatus =
  (typeof OrderStatusInputStatus)[keyof typeof OrderStatusInputStatus];

export const OrderStatusInputStatus = {
  pending: "pending",
  processing: "processing",
  shipped: "shipped",
  delivered: "delivered",
  cancelled: "cancelled",
} as const;

export interface OrderStatusInput {
  status: OrderStatusInputStatus;
}

export type AdminStatsOrdersByStatusItem = {
  status: string;
  count: number;
};

export interface AdminStats {
  totalRevenue: number;
  totalOrders: number;
  totalProducts: number;
  totalUsers: number;
  recentOrders: Order[];
  ordersByStatus: AdminStatsOrdersByStatusItem[];
}

export type GetProductsParams = {
  search?: string;
  category?: string;
  minPrice?: number;
  maxPrice?: number;
};