# lib/api-zod/src/generated/types/adminStats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * E-Commerce API
 * OpenAPI spec version: 0.1.0
 */
import type { AdminStatsOrdersByStatusItem } from "./adminStatsOrdersByStatusItem";
import type { Order } from "./order";

export interface AdminStats {
  totalRevenue: number;
  totalOrders: number;
  totalProducts: number;
  totalUsers: number;
  recentOrders: Order[];
  ordersByStatus: AdminStatsOrdersByStatusItem[];
}