# lib/api-zod/src/generated/types/orderInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * E-Commerce API
 * OpenAPI spec version: 0.1.0
 */
import type { OrderInputItemsItem } from "./orderInputItemsItem";

export interface OrderInput {
  items: OrderInputItemsItem[];
  shippingAddress: string;
}