# lib/api-zod/src/generated/types/productInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * E-Commerce API
 * OpenAPI spec version: 0.1.0
 */

export interface ProductInput {
  name: string;
  description: string;
  price: number;
  image: string;
  category: string;
  countInStock: number;
}